Rename files to prevent accidental library mismatch
